# src/api/swagger.json
{
    "openapi": "3.0.0",
    "servers": [],
    "info": {
        "title": "",
        "description": "",
        "version": ""
    },
    "tags": [{
            "name": "Info"
        },
        {
            "name": "User"
        },
        {
            "name": "Image"
        }
    ],
    "paths": {
        "/": {
            "get": {
                "tags": [
                    "Info"
                ],
                "summary": "Show API information",
                "description": "This is a public route",
                "operationId": "showApiInfo",
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "List all users",
                "description": "Returns all users",
                "operationId": "FindAllUsers",
                "security": [{
                    "basicAuth": []
                }],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create new user",
                "description": "",
                "operationId": "CreateUser",
                "security": [{
                    "basicAuth": []
                }],
                "responses": {
                    "201": {
                        "description": "created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "$ref": "#/components/requestBodies/NewUser"
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user",
                "description": "Returns the given user",
                "operationId": "FindUser",
                "security": [{
                    "basicAuth": []
                }],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of user to return",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "format": "uuid"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user",
                "description": "Updates the given user",
                "operationId": "UpdateUser",
                "security": [{
                    "basicAuth": []
                }],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of user to update",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "format": "uuid"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "$ref": "#/components/requestBodies/NewUser"
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user",
                "description": "Removes the given user",
                "operationId": "DeleteUser",
                "security": [{
                    "basicAuth": []
                }],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of user to delete",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "format": "uuid"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/images": {
            "get": {
                "tags": [
                    "Image"
                ],
                "summary": "List all images",
                "description": "Returns all images",
                "operationId": "FindAllImages",
                "security": [{
                    "basicAuth": []
                }],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Image"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Image"
                ],
                "summary": "Create new image",
                "description": "",
                "operationId": "CreateImage",
                "security": [{
                    "basicAuth": []
                }],
                "responses": {
                    "201": {
                        "description": "created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Image"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewImage"
                            }
                        }
                    },
                    "description": "User object that needs to be added to the database",
                    "required": true
                }
            }
        },
        "/images/{id}": {
            "get": {
                "tags": [
                    "Image"
                ],
                "summary": "Get image",
                "description": "Returns the given image",
                "operationId": "FindImage",
                "security": [{
                    "basicAuth": []
                }],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of image to return",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "format": "uuid"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Image"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Image"
                ],
                "summary": "Update image",
                "description": "Updates the given image",
                "operationId": "UpdateImage",
                "security": [{
                    "basicAuth": []
                }],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of image to update",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "format": "uuid"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Image"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewImage"
                            }
                        }
                    },
                    "description": "Image object that needs to be added to the database",
                    "required": true
                }
            },
            "delete": {
                "tags": [
                    "Image"
                ],
                "summary": "Delete image",
                "description": "Removes the given image",
                "operationId": "DeleteImage",
                "security": [{
                    "basicAuth": []
                }],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of image to delete",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "format": "uuid"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "requestBodies": {
            "NewUser": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/NewUser"
                        }
                    }
                },
                "description": "User object that needs to be added to the database",
                "required": true
            }
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        },
        "schemas": {
            "NewUser": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                }
            },
            "NewImage": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "age": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "userId": {
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
            "User": {
                "title": "User",
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "Image": {
                "title": "Image",
                "allOf": [{
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "name": {
                            "type": "string"
                        },
                        "age": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "user": {
                            "$ref": "#/components/schemas/User"
                        },
                        "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                        },
                        "createdAt": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                }]
            }
        }
    }
}